Moved host tools, fixed mentor registration, deleted old models, updated imports
