+++ registration/backend/app/src/database/starter.py
@@ -33,3 +33,4 @@ def get_db():
         yield db
     finally:
         db.close()
+
