+++ registration/backend/app/src/helpers/dbtools.py
@@ -88,6 +88,7 @@ def create_user(db: Session, user: Union[mentor.Mentor, attendee.Attendee],
             last_name=user.last_name,
             email=user.email,
             phone_number=user.phone_number,
+            occupation=user.occupation,
             country=user.country,
             address_line_one=user.address_line_one,
             address_line_two=user.address_line_two,
